% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_faceted_clusters}
\alias{plot_faceted_clusters}
\title{Plot clusters by facet}
\usage{
plot_faceted_clusters(spe, annotationsName)
}
\arguments{
\item{spe}{A SpatialExperiment or SpatialFeatureExperiment object}

\item{annotationsName}{the name of the \code{colData(spe)} column that stores the annotations to plot}
}
\value{
A list of \code{ggplot} objects with each item in the list being a plot of a level in \code{annotationsName}
}
\description{
Plot clusters by facet
}
