% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_NMF.R
\name{run_nmf}
\alias{run_nmf}
\title{Run NMF on a dataset}
\usage{
run_nmf(data, assay, k)
}
\arguments{
\item{data}{SingleCellExperiment or SpatialExperiment object}

\item{assay}{string indicating the assay to run NMF on}

\item{k}{integer indicating the number of factors for NMF}
}
\description{
Run NMF on a dataset
}
