% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_helpers.R
\name{run_rank_determination_nmf}
\alias{run_rank_determination_nmf}
\title{run_rank_determination_nmf}
\usage{
run_rank_determination_nmf(data, assay)
}
\arguments{
\item{data}{A SingleCellExperiment or SpatialExperiment object}

\item{assay}{string indicating the assay to run NMF on}
}
\value{
a NMF model object
}
\description{
run_rank_determination_nmf
}
