% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_labels.R
\name{smoother}
\alias{smoother}
\title{k-Nearest neighbor cluster label smoothing with level-specific thresholding.}
\usage{
smoother(
  labels_curr,
  locs,
  k = 10L,
  props = NULL,
  max_iter = 10,
  verbose = TRUE
)
}
\arguments{
\item{labels_curr}{a vector of labels for each observation}

\item{locs}{A dataframe of spatial coordinates corresponding to the observations}

\item{k}{An integer scalar specifying number of neighbors for smoothing.}

\item{props}{A named list of numeric scalars \eqn{\in [0,1]} specifying a label
proportions threshold for each level in the labels. If the fraction of neighbors with a certain label
exceeds this proportion, change the label of the current sample
(default: 0.5 for each level).The names of the list should correspond to the levels in the label.}

\item{max_iter}{An integer scalar specifying the max number of smoothing
iterations. Set to -1 for smoothing to convergence.}

\item{verbose}{A logical scalar representing verbosity}
}
\description{
k-Nearest neighbor cluster label smoothing with level-specific thresholding.
}
