% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_utils.R
\name{compute_factor_correlations}
\alias{compute_factor_correlations}
\title{Compute correlation of each NMF factor with annotations in the source dataset}
\usage{
compute_factor_correlations(factors, cor_variable)
}
\arguments{
\item{factors}{A Data.Frame of NMF factors}

\item{cor_variable}{A vector of per-cell annotations or discrete technical levels for your source dataset. For example, cell-type labels or sample ID.}
}
\value{
A NMF factor by unique annotations matrix of correlations between each factor and each unique annotation.
}
\description{
Compute correlation of each NMF factor with annotations in the source dataset
}
