% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_helpers.R
\name{run_nmf}
\alias{run_nmf}
\title{Run NMF on a dataset}
\usage{
run_nmf(data, assay, k = NULL, seed = 1237, ...)
}
\arguments{
\item{data}{SingleCellExperiment or SpatialExperiment object}

\item{assay}{string indicating the assay to run NMF on}

\item{k}{integer indicating the number of factors for NMF}

\item{seed}{a random seed}

\item{...}{additional arguments passed to singlet::run_nmf or singlet::RunNMF}
}
\value{
NMF model object
}
\description{
Run NMF on a dataset
}
