% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_labels.R
\name{transfer_labels}
\alias{transfer_labels}
\title{Transfer labels from the source dataset to the target dataset}
\usage{
transfer_labels(source, target)
}
\arguments{
\item{source}{A SingleCellExperiment or SpatialExperiment object with cell-type or spatial domain labels}

\item{target}{A SingleCellExperiment or SpatialExperiment object to transfer labels into}
}
\value{
A SingleCellExperiment or SpatialExperiment objec
}
\description{
Transfer labels from the source dataset to the target dataset
}
